4.0
Welcome in the CPS world! Plugin is rebranded for Cherry Pick Studios brand. This release added a new popup image option to show an image or logo on the popup. There is also a new option to choose the style of the popup window and buttons. The CPS SDK is added to the plugin.

NEW
- New plugin name and brand.
- Image field for the popup.
- Three new style: default, almost flat, gradient.

ENHACEMENTS
- UIkit framework added to the plugin.
- UIkit is updated to 2.27.5 version.
- Javascript functions got unique names to avoid conflicts.
- The admin function has been switched to CPS SDK.
- CSS and JS files are moved to assets folder.
- UIkit framework moved to vendors folder.
- Admin menu moved under the CPS Plugins main menu.

OTHER
- Tested with WordPress 5.2 version.
- Added minimum PHP version requirement.