5.0
The Age Verification Premium is here! There are a lot of new and fantastic features to manage and design your Age Verification popup. Please consider to buy the Premium version not only for the new features, but for supporting our work.

IMPORTANT! A lot of new codes are added and modified. The logic of the design is changed under the hood. There are options also, that were moved to Premium only features. These options can not be changed in the free version, but you can keep your current options without losing it.

NEW

- New premium options added to the admin.
- PRO - Add Image to the header.
- PRO - Add background image to the popup.
- PRO - A lot of cool design features added, like Dark mode, Vertical centering, Large modifier, etc.
- PRO - You can enable closing the popup with a button, ESC or clicking on the background.
- PRO - Give delay to the popup.

ENHANCEMENTS

- Code optimization from the ground up.
- A lot of security hardening for input elements.
- Validation enhancements.
- Added new field for WPML translation.

OTHER

- Changes in the description.
- Settings page sidebar changes.
- Welcome notice added.